# Empresa-OpenApi.yaml
openapi: 3.0.0
info:
  title: Cuentas
  version: '1.0'
  description: Operaciones con Cuentas
  contact:
    name: William
    url: 'http://wquimis.com'
    email: wquimis@gmail.com
  termsOfService: 'http://wquimis.com'
servers:
  - url: 'http://localhost:3000'
...
paths:
...
  /movimientos:
    parameters: []
    post:
      summary: Crear Movimiento
      operationId: crearmovimientos
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movimiento'
              examples:
                movimiento:
                  value:
                    idMovimiento: 0
                    idCuenta: 0
                    fecha: '2019-08-24T14:15:22Z'
                    tipoMovimiento: A
                    valor: 0
                    saldo: 0
                    movDescripcion: Ninguno
                    Estado: R
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - IdError: string
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - IdError: string
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movimiento'
            examples:
              movimiento:
                value:
                  idMovimiento: 0
                  idCuenta: 0
                  fecha: '2019-08-24T14:15:22Z'
                  tipoMovimiento: A
                  valor: 0
                  saldo: 0
                  movDescripcion: Ninguno
                  Estado: R
        description: Devuelve el objeto creado con su ID
      description: Creacion de Movimientos
      tags:
        - Movimientos
    get:
      summary: Consultar Movimientos
      operationId: obtenermovimientos
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movimiento'
              examples:
                movimiento:
                  value:
                    - idMovimiento: 0
                      idCuenta: 0
                      fecha: '2019-08-24T14:15:22Z'
                      tipoMovimiento: A
                      valor: 0
                      saldo: 0
                      movDescripcion: Ninguno
                      Estado: R
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
      description: Consulta de Movimientos
      parameters:
        - schema:
            type: integer
          in: query
          name: idmovimiento
          description: Identificador de movimiento
      tags:
        - Movimientos
  /movimientos/reporte:
    parameters: []
    get:
      summary: Consulta Movimientos Por Fecha
      operationId: obtenermovimientosreporte
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MovimientoReporte'
              examples:
                reporte:
                  value:
                    - fecha: '2019-08-24T14:15:22Z'
                      cliente: string
                      numeroCuenta: string
                      tipoCuenta: string
                      saldoInicial: 0
                      estadoMovimiento: string
                      movimiento: 0
                      saldoDisponible: 0
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
      description: Consulta de Movimientos por fechas
      parameters:
        - schema:
            type: string
            format: date-time
          in: query
          name: inicio
          description: Fecha de inicio
          required: true
        - schema:
            type: string
            format: date-time
          in: query
          name: fin
          description: Fecha de fin
          required: true
        - schema:
            type: integer
          in: query
          name: idcliente
          description: Identificador del cliente
          required: true
      tags:
        - Movimientos
  '/movimientos/{idmovimiento}':
    parameters:
      - schema:
          type: integer
        name: idmovimiento
        in: path
        required: true
    put:
      summary: Actualizar Movimiento
      operationId: actualizarmovimiento
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movimiento'
              examples:
                movimiento:
                  value:
                    idMovimiento: 0
                    idCuenta: 0
                    fecha: '2019-08-24T14:15:22Z'
                    tipoMovimiento: A
                    valor: 0
                    saldo: 0
                    movDescripcion: Ninguno
                    Estado: R
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
      description: Actualizar movimiento
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Movimiento'
            examples:
              movimiento:
                value:
                  idMovimiento: 0
                  idCuenta: 0
                  fecha: '2019-08-24T14:15:22Z'
                  tipoMovimiento: A
                  valor: 0
                  saldo: 0
                  movDescripcion: Ninguno
                  Estado: R
        description: SOlicitud de movimiento
      tags:
        - Movimientos
    delete:
      summary: Eliminar Movimiento
      operationId: eliminarmovimiento
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
              examples:
                errores:
                  value:
                    - idError: 0
                      mensajeTecnico: string
                      mensajeUsuario: string
                      nivelError: string
      description: Elimina un movimiento con IdMovimiento
      tags:
        - Movimientos
components:
  schemas:
...
    Error:
      title: Error
      type: object
      x-examples: {}
      description: Modelos de errores
      properties:
        idError:
          type: integer
        mensajeTecnico:
          type: string
        mensajeUsuario:
          type: string
        nivelError:
          type: string
...
    Movimiento:
      title: Movimiento
      x-stoplight:
        id: hen7kygnnd3da
      type: object
      description: Modelo Movimientos financieros
      properties:
        idMovimiento:
          type: integer
        idCuenta:
          type: integer
        fecha:
          type: string
          format: date-time
        tipoMovimiento:
          type: string
          maxLength: 1
          default: A
        valor:
          type: number
          minimum: -999999999999.999
          maximum: 999999999999.999
          default: 0
          format: double
        saldo:
          type: number
          default: 0
          format: double
          minimum: -999999999999.999
          maximum: 999999999999.999
        movDescripcion:
          type: string
          maxLength: 200
          default: Ninguno
        Estado:
          type: string
          default: R
          maxLength: 1
      required:
        - idCuenta
        - tipoMovimiento
        - valor
        - Estado
    MovimientoReporte:
      type: object
      x-examples: {}
      properties:
        fecha:
          type: string
          format: date-time
        cliente:
          type: string
        numeroCuenta:
          type: string
        tipoCuenta:
          type: string
        saldoInicial:
          type: number
          format: double
        estadoMovimiento:
          type: string
        movimiento:
          type: number
          format: double
        saldoDisponible:
          type: number
          format: double